# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Stackr API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/apps": {
      "get": {
        "summary": "List apps (scope: apps:read)",
        "operationId": "listApps",
        "parameters": [
          {
            "name": "stack",
            "in": "query",
            "description": "filter to a stack id",
            "schema": {
              "type": "string",
              "description": "filter to a stack id"
            }
          },
          {
            "name": "env",
            "in": "query",
            "description": "filter to an environment id",
            "schema": {
              "type": "string",
              "description": "filter to an environment id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/V1AppOut"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
    "/api/v1/apps/{id}": {
      "delete": {
        "summary": "Delete app (scope: apps:write)",
        "operationId": "deleteApp",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      },
      "get": {
        "summary": "Get app (scope: apps:read)",
        "operationId": "getApp",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1AppOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      },
      "patch": {
        "summary": "Update app (scope: apps:write)",
        "operationId": "updateApp",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V1AppPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1AppOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
...
    "/api/v1/deployments/{id}": {
      "get": {
        "summary": "Get deployment (scope: apps:read)",
        "operationId": "getDeployment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1DeploymentOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
...
    "/api/v1/stacks/{id}/apps": {
      "post": {
        "summary": "Create app (scope: apps:write)",
        "operationId": "createApp",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V1AppIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1AppOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ErrBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "StackconfBackupConf": {
        "type": "object",
        "properties": {
          "dest": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean",
            "nullable": true
          },
          "keep": {
            "type": "integer"
          },
          "mode": {
            "type": "string"
          },
          "schedule": {
            "type": "string"
          },
          "tz": {
            "type": "string"
          }
        }
      },
      "StackconfBuildConf": {
        "type": "object",
        "properties": {
          "context": {
            "type": "string"
          },
          "dockerfile": {
            "type": "string"
          }
        }
      },
...
      "StackconfEnvMap": {
        "type": "object",
        "additionalProperties": {
          "type": "string"
        }
      },
      "StackconfFileList": {
        "type": "array",
        "items": {
          "type": "string"
        }
      },
...
      "V1AppOut": {
        "type": "object",
        "properties": {
          "env_id": {
            "type": "string"
          },
          "git_branch": {
            "type": "string"
          },
          "git_url": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "image_digest": {
            "type": "string"
          },
          "latest_digest": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "source_type": {
            "type": "string"
          },
          "stack_id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "update_policy": {
            "type": "string"
          },
          "wait_for_ci": {
            "type": "boolean"
          }
        }
      },
      "V1AppPatch": {
        "type": "object",
        "properties": {
          "allow_overlap": {
            "type": "boolean"
          },
          "attach": {
            "type": "string"
          },
          "backup": {
            "$ref": "#/components/schemas/StackconfBackupConf"
          },
          "basic_auth_password": {
            "type": "string"
          },
          "basic_auth_user": {
            "type": "string"
          },
          "branch": {
            "type": "string"
          },
          "build": {
            "$ref": "#/components/schemas/StackconfBuildConf"
          },
          "build_args": {
            "type": "string"
          },
          "command": {
            "type": "string"
          },
          "connector": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "devices": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "domains": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StackconfDomainConf"
            }
          },
          "engine": {
            "type": "string"
          },
          "env": {
            "$ref": "#/components/schemas/StackconfEnvMap"
          },
          "external_port": {
            "type": "integer"
          },
          "files": {
            "$ref": "#/components/schemas/StackconfFileList"
          },
          "from": {
            "type": "string"
          },
          "git_branch": {
            "type": "string"
          },
          "git_url": {
            "type": "string"
          },
          "healthcheck": {
            "type": "string"
          },
          "healthcheck_interval": {
            "type": "integer"
          },
          "healthcheck_retries": {
            "type": "integer"
          },
          "healthcheck_start_period": {
            "type": "integer"
          },
          "healthcheck_timeout": {
            "type": "integer"
          },
          "image": {
            "type": "string"
          },
          "limits": {
            "$ref": "#/components/schemas/StackconfLimitsConf"
          },
          "name": {
            "type": "string"
          },
          "node_group": {
            "type": "string"
          },
          "on_remove": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "port": {
            "type": "integer"
          },
          "privileged": {
            "type": "boolean"
          },
          "public": {
            "type": "boolean"
          },
          "published_ports": {
            "type": "string"
          },
          "replicas": {
            "type": "integer"
          },
          "restart": {
            "type": "string"
          },
          "run_on_deploy": {
            "type": "boolean"
          },
          "schedule": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "security_headers": {
            "type": "boolean"
          },
          "shm_size_mb": {
            "type": "integer"
          },
          "storage": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "timeout_minutes": {
            "type": "integer"
          },
          "traefik_override": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "update_policy": {
            "type": "string"
          },
          "user": {
            "type": "string"
          },
          "volumes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "wait_for_ci": {
            "type": "boolean"
          },
          "watch_paths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "V1DeploymentOut": {
        "type": "object",
        "properties": {
          "app_id": {
            "type": "string"
          },
          "commit_sha": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "image_tag": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "trigger": {
            "type": "string"
          }
        }
      },
...
      "V1ErrBody": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/V1ErrDetail"
          }
        }
      },
      "V1ErrDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "x-api-key",
        "in": "header",
        "description": "API key"
      }
    }
  }
}